+++ tools/ncurses/tools/axis.h
@@ -17,3 +17,5 @@ struct x_axis_value {
 struct x_axis_range {
 	struct timeval start, end;
 };
+
+const char *x_axis_range_str(struct x_axis_range *v, char buf[64]);