+++ tools/ncurses/tools/plot.c
@@ -133,12 +133,6 @@ void __plot_warning(const struct plot *p, char *fmt, ...)
 	attroff(colors[C_RED] | A_BOLD);
 }
 
-static const char *x_axis_tv_str(struct timeval *tv, char buf[32])
-{
-	strftime(buf, 32, "%T", localtime(&tv->tv_sec));
-	return buf;
-}
-
 /**
  * @start: start point of line.
  * @len: number of value to plot.
@@ -268,7 +262,7 @@ static void __paint_line(struct plot *p, const struct lgroup *lg,
 		if ((ivs - 1) % 10 == 0) {
 			char buf[32];
 			mvprintw(p->height - p->bnd.bottom + 1, w, "%s",
-				 x_axis_tv_str(&v->x_axis.tv, buf));
+				 timeval_str(&v->x_axis.tv, buf));
 		}
 
 		/* set y axis */
@@ -418,13 +412,12 @@ void __plot_debug_llabel(const struct lgroup *lg, int height)
 			mvprintw(i + height, p->bnd.left + 1, "%d: %s: %ld",
 				 ln->id, ln->name, ln->count);
 		else {
-			char buf1[32], buf2[32];
+			char buf[64];
 
 			mvprintw(i + height, p->bnd.left + 1,
-				 "%d: %s: %ld %f x(%s~%s) y(%lf~%lf)", ln->id,
+				 "%d: %s: %ld %f x(%s) y(%lf~%lf)", ln->id,
 				 ln->name, ln->count, ln->tail->v,
-				 x_axis_tv_str(&ln->x_axis_range.start, buf1),
-				 x_axis_tv_str(&ln->x_axis_range.end, buf2),
+				 x_axis_range_str(&ln->x_axis_range, buf),
 				 ln->min->v, ln->max->v);
 		}
 		attroff(color);