+++ tools/ncurses/tools/utils.h
@@ -4,7 +4,9 @@
 #include <sys/time.h>
 
 unsigned long usecs(void);
+const char *timeval_str(struct timeval *tv, char buf[32]);
 struct timeval max_timeval(struct timeval *tv1, struct timeval *tv2);
+struct timeval diff_timeval(struct timeval *tv1, struct timeval *tv2);
 
 unsigned long str2nsecs(const char *str);